Concise commit message:
```text
refactor: update Hilt ViewModel imports and refine Compose UI components

- Update `hiltViewModel` imports from `androidx.hilt.navigation.compose` to `androidx.hilt.lifecycle.viewmodel.compose`.
- Replace `centerAlignedTopAppBarColors` with `topAppBarColors` in security and hub screens.
- Migrate to `AutoMirrored` version of the back arrow icon in `NotchBarGateCard`.
- Add `@Suppress("UNCHECKED_CAST")` to Flow transformation logic in `CustomizationViewModel` to silence lint warnings.
- Simplify metadata assignment in `TrinityRepository` by removing redundant `toString()` mapping.
- Clean up unnecessary `.toFloat()` conversions in `DataVisualizationBackground` arithmetic.
```

+++ app/src/main/java/dev/aurakai/auraframefx/domains/aura/ui/components/backgrounds/DataVisualizationBackground.kt
@@ -119,8 +119,8 @@ fun DataVisualizationBackground(
                 val radius = (0.3f + 0.7f * progress) * maxRadius * (1 + noise * 0.2f)
 
                 Offset(
-                    x = center.x + radius * cos(angle + noise * 0.2f).toFloat(),
-                    y = center.y + radius * sin(angle + noise * 0.2f).toFloat()
+                    x = center.x + radius * cos(angle + noise * 0.2f),
+                    y = center.y + radius * sin(angle + noise * 0.2f)
                 )
             }
